+++ src/main/java/duke/Launcher.java
@@ -1,9 +1,10 @@
 package duke.application;
 
+import duke.Duke;
 import javafx.application.Application;
 
 public class Launcher {
     public static void main(String[] args) {
-        Application.launch(DukeApplication.class, args);
+        Application.launch(Duke.class, args);
     }
 }
